# docs/swagger.yaml
definitions:
...
  models.AccountBreakdown:
    properties:
      account_id:
        type: string
      account_name:
        type: string
      total:
        type: integer
      type:
        type: string
    type: object
...
  models.CategoryBreakdown:
    properties:
      category_id:
        type: string
      category_name:
        type: string
      total:
        type: integer
      type:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.ReportByAccount:
    properties:
      accounts:
        items:
          $ref: '#/definitions/models.AccountBreakdown'
        type: array
      period_end:
        type: string
      period_start:
        type: string
    type: object
  models.ReportByCategory:
    properties:
      categories:
        items:
          $ref: '#/definitions/models.CategoryBreakdown'
        type: array
      period_end:
        type: string
      period_start:
        type: string
    type: object
  models.ReportSummary:
    properties:
      base_currency:
        type: string
      period_end:
        type: string
      period_start:
        type: string
      total_balance:
        type: integer
      total_expenses:
        type: integer
      total_income:
        type: integer
      total_transfers:
        type: integer
    type: object
  models.ReportSummaryResponse:
    properties:
      data:
        $ref: '#/definitions/models.ReportSummary'
    type: object
  models.ReportTrends:
    properties:
      data_points:
        items:
          $ref: '#/definitions/models.TrendDataPoint'
        type: array
      period_end:
        type: string
      period_start:
        type: string
    type: object
...
  models.TrendDataPoint:
    properties:
      date:
        type: string
      total:
        type: integer
      type:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /reports/by-account:
    get:
      consumes:
      - application/json
      description: Get aggregated spending data grouped by account
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ReportByAccount'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get spending breakdown by account
      tags:
      - reports
  /reports/by-category:
    get:
      consumes:
      - application/json
      description: Get aggregated spending data grouped by category
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ReportByCategory'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get spending breakdown by category
      tags:
      - reports
  /reports/summary:
    get:
      consumes:
      - application/json
      description: Get aggregated financial data including totals by type and category
        breakdown
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ReportSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get financial report summary
      tags:
      - reports
  /reports/trends:
    get:
      consumes:
      - application/json
      description: Get time-series data for charting
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ReportTrends'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get time-series trend data
      tags:
      - reports
...
swagger: "2.0"