# internal/interface/handler/report.go
package handler

import (
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/dimasbaguspm/penster/internal/application/service"
	"github.com/dimasbaguspm/penster/internal/interface/dto"
	"github.com/dimasbaguspm/penster/pkg/models"
	"github.com/dimasbaguspm/penster/pkg/observability"
)

type ReportHandler struct {
	svc *service.ReportService
}

func NewReportHandler(svc *service.ReportService) *ReportHandler {
	return &ReportHandler{svc: svc}
}

// Summary handles GET /reports/summary
// @Summary Get financial report summary
// @Description Get aggregated financial data including totals by type and category breakdown
// @Tags reports
// @Accept json
// @Produce json
// @Param start_date query string true "Start date (YYYY-MM-DD)"
// @Param end_date query string true "End date (YYYY-MM-DD)"
// @Success 200 {object} models.ReportSummaryResponse
// @Failure 400 {object} models.ErrorResponse
// @Failure 500 {object} models.ErrorResponse
// @Router /reports/summary [get]
func (h *ReportHandler) Summary(w http.ResponseWriter, r *http.Request) {
	log := observability.NewLogger(r.Context(), "http", "report")
	ctx, span := observability.StartHandlerSpan(log.Context(), "Report", "Summary")
	defer span.End()

	startDate, endDate, err := dto.ParseReportParams(r)
	if err != nil {
		log.Warn("invalid report params", "error", err)
		h.writeError(w, http.StatusBadRequest, err.Error())
		return
	}

	log.Info("getting summary report", "start_date", startDate.Format("2006-01-02"), "end_date", endDate.Format("2006-01-02"))
	summary, err := h.svc.GetSummary(ctx, startDate.Format("2006-01-02"), endDate.Format("2006-01-02"))
	if err != nil {
		log.Error("failed to get summary report", "error", err)
		h.writeError(w, http.StatusInternalServerError, err.Error())
		return
	}

	log.Info("summary report retrieved")
	h.writeJSON(w, http.StatusOK, models.ReportSummaryResponse{Data: *summary})
}

// ByAccount handles GET /reports/by-account
// @Summary Get spending breakdown by account
// @Description Get aggregated spending data grouped by account
// @Tags reports
// @Accept json
// @Produce json
// @Param start_date query string true "Start date (YYYY-MM-DD)"
// @Param end_date query string true "End date (YYYY-MM-DD)"
// @Success 200 {object} models.ReportByAccount
// @Failure 400 {object} models.ErrorResponse
// @Failure 500 {object} models.ErrorResponse
// @Router /reports/by-account [get]
func (h *ReportHandler) ByAccount(w http.ResponseWriter, r *http.Request) {
	log := observability.NewLogger(r.Context(), "http", "report")
	ctx, span := observability.StartHandlerSpan(log.Context(), "Report", "ByAccount")
	defer span.End()

	startDate, endDate, err := dto.ParseReportParams(r)
	if err != nil {
		log.Warn("invalid report params", "error", err)
		h.writeError(w, http.StatusBadRequest, err.Error())
		return
	}

	log.Info("getting report by account", "start_date", startDate.Format("2006-01-02"), "end_date", endDate.Format("2006-01-02"))
	report, err := h.svc.GetByAccount(ctx, startDate.Format("2006-01-02"), endDate.Format("2006-01-02"))
	if err != nil {
		log.Error("failed to get report by account", "error", err)
		h.writeError(w, http.StatusInternalServerError, err.Error())
		return
	}

	log.Info("report by account retrieved")
	h.writeJSON(w, http.StatusOK, report)
}

// ByCategory handles GET /reports/by-category
// @Summary Get spending breakdown by category
// @Description Get aggregated spending data grouped by category
// @Tags reports
// @Accept json
// @Produce json
// @Param start_date query string true "Start date (YYYY-MM-DD)"
// @Param end_date query string true "End date (YYYY-MM-DD)"
// @Success 200 {object} models.ReportByCategory
// @Failure 400 {object} models.ErrorResponse
// @Failure 500 {object} models.ErrorResponse
// @Router /reports/by-category [get]
func (h *ReportHandler) ByCategory(w http.ResponseWriter, r *http.Request) {
	log := observability.NewLogger(r.Context(), "http", "report")
	ctx, span := observability.StartHandlerSpan(log.Context(), "Report", "ByCategory")
	defer span.End()

	startDate, endDate, err := dto.ParseReportParams(r)
	if err != nil {
		log.Warn("invalid report params", "error", err)
		h.writeError(w, http.StatusBadRequest, err.Error())
		return
	}

	log.Info("getting report by category", "start_date", startDate.Format("2006-01-02"), "end_date", endDate.Format("2006-01-02"))
	report, err := h.svc.GetByCategory(ctx, startDate.Format("2006-01-02"), endDate.Format("2006-01-02"))
	if err != nil {
		log.Error("failed to get report by category", "error", err)
		h.writeError(w, http.StatusInternalServerError, fmt.Sprintf("ByCategory error: %v", err))
		return
	}

	log.Info("report by category retrieved")
	h.writeJSON(w, http.StatusOK, report)
}

// Trends handles GET /reports/trends
// @Summary Get time-series trend data
// @Description Get time-series data for charting
// @Tags reports
// @Accept json
// @Produce json
// @Param start_date query string true "Start date (YYYY-MM-DD)"
// @Param end_date query string true "End date (YYYY-MM-DD)"
// @Success 200 {object} models.ReportTrends
// @Failure 400 {object} models.ErrorResponse
// @Failure 500 {object} models.ErrorResponse
// @Router /reports/trends [get]
func (h *ReportHandler) Trends(w http.ResponseWriter, r *http.Request) {
	log := observability.NewLogger(r.Context(), "http", "report")
	ctx, span := observability.StartHandlerSpan(log.Context(), "Report", "Trends")
	defer span.End()

	startDate, endDate, err := dto.ParseReportParams(r)
	if err != nil {
		log.Warn("invalid report params", "error", err)
		h.writeError(w, http.StatusBadRequest, err.Error())
		return
	}

	log.Info("getting trends report", "start_date", startDate.Format("2006-01-02"), "end_date", endDate.Format("2006-01-02"))
	report, err := h.svc.GetTrends(ctx, startDate.Format("2006-01-02"), endDate.Format("2006-01-02"))
	if err != nil {
		log.Error("failed to get trends report", "error", err)
		h.writeError(w, http.StatusInternalServerError, err.Error())
		return
	}

	log.Info("trends report retrieved")
	h.writeJSON(w, http.StatusOK, report)
}

func (h *ReportHandler) writeJSON(w http.ResponseWriter, status int, data any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	json.NewEncoder(w).Encode(data)
}

func (h *ReportHandler) writeError(w http.ResponseWriter, status int, message string) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	json.NewEncoder(w).Encode(models.ErrorResponse{Error: message})
}